refactor: moved lexer tests to test/
