+++ src/lib.rs
@@ -1,0 +1,10 @@
+pub mod ast;
+pub mod builtin;
+pub mod env;
+pub mod eval;
+pub mod lexer;
+pub mod object;
+pub mod parser;
+
+pub use lexer::Lexer;
+pub use parser::Parser;
